Step 3 of the switch to support configuration with NPTL: Change "configure.in".
The branches for Tru64 (ex: OSF/1) and Linux are separated,
and for Linux 'getconf GNU_LIBPTHREAD_VERSION' is used to tell between Linuxthreads and NPTL.

--BZR--
revision-id: sp1r-joerg@mysql.com-20050422184905-57565
property-file-info: ld7:file_id65:sp1f-configure.in-19700101030959-mgdpoxtnh2ewmvusvfpkreuhwvffkcjw7:message1555:Step 3 of the switch to support configuration with NPTL:
property-file-info: Change "configure.in".
property-file-info: 
property-file-info: 1) Separate the branches checking for thread libraries on OSF/1 (now: Tru64) and on Linux,
property-file-info:    as these are different things. In the Tru64 branch, remove the obsolete part for OSF/1 3.2,
property-file-info:    because that is gone since long (about a decade?); we are at 5 now. Adapt the comment.
property-file-info: 
property-file-info: 2) For Linux, the thread libraries are "Linuxthreads" (older) and "NPTL" (newer).
property-file-info:    Until now, we cared only about "Linuxthreads", and we did so by checking a comment in 
property-file-info:    the header file '/usr/include/pthread.h'.
property-file-info:    With NPTL as an alternative or only thread package, this is not secure any more,
property-file-info:    there are distributions where the header file still contains that comment but the lib is NPTL.
property-file-info:    For all current distributions, the way to check is the 'getconf' command.
property-file-info: 
property-file-info:    On old distributions, 'getconf' exists but does not understand the 'GNU_LIBPTHREAD_VERSION'
property-file-info:    parameter. It is unlikely that such an old system should have NPTL, so the header file check is
property-file-info:    sufficient if the getconf inquiry fails.
property-file-info: 
property-file-info:    For both Linuxthreads and NPTL, '-lpthread' is the linker option to use.
property-file-info: 
property-file-info: 3) To detect problems as early as possible, on Linux a link test using '-lpthread' is done.
property-file-info:    The code is copied from a later check where the thread functions are searched in '-lc',
property-file-info:    '-lpthread', and '-lpthreads', but this later search would not be executed after the lib
property-file-info:    was already determined. So if the test is not done here, it will be detected only during
property-file-info:    the real build. This is too late.
property-file-info: 4:path12:configure.inee
property-sp1-file-info: ld9:commit_id49:joerg@mysql.com|configure.in|20050422184901|307527:file_id53:BK|configure.in|19700101030959|00052|606b4ee8690d8c62ee
testament3-sha1: 9f5cacaf8160de98197e2666cf3596127ec049b4