BUG#11132 Connections stuck in CLOSE_WAIT
This fixes the problem in 5.0 in the correct way.

The previous patch (for 4.1) is only valid for 4.1 and is done at the right time
for 4.1.

--BZR--
revision-id: sp1r-stewart@mysql.com-20050616081744-19213
property-file-info: ld7:file_id61:sp1f-main.cpp-20040414082422-ifhdrjyod62azjl4peccmcmjokvsy3na7:message556:The ndbd angel process keeps the mgm connection open as it retains a reference
property-file-info: to the socket.
property-file-info: 
property-file-info: this means that when the mgm server is killed, the socket stays in CLOSE_WAIT 
property-file-info: (the ndb process will close() it okay in 5.0, as we reused it for a transporter
property-file-info: but this isn't enough as the angel process still has a reference to that socket).
property-file-info: 
property-file-info: so, by adding in a theConfig->closeConfiguration(); in the angel process, the ndb
property-file-info: process is left with the only reference to the socket - which it can close whenever
property-file-info: it likes (namely when the connection is terminated).
property-file-info: 4:path23:ndb/src/kernel/main.cppee
property-sp1-file-info: ld9:commit_id62:stewart@mysql.com|ndb/src/kernel/main.cpp|20050616081737|379417:file_id93:magnus@neptunus.(none)|ndb/src/kernel/ndb-main/Main.cpp|20040414082422|15930|d9731c8aace8895aee
testament3-sha1: db88d5e67a2f25c7fab8abf1eee55853534bc4f7