Fix for bug #12456: Cursor shows incorrect data - DML does not affect, probably caching. Fixed test results.
This patch implements a high-granularity read view to be used with
cursors. In this high-granularity consistent read view modifications 
done by the creating transaction after the cursor is created or 
future transactions are not visible. But those modifications that 
transaction did before the cursor was created are visible.
--BZR--
revision-id: svn-v4:cee13dc7-1704-0410-992b-c9b4543f1246:trunk:259
testament3-sha1: fc5125dc42a9ac6d3290ce52b41a614909848be7