innodb.test, innodb.result: Add test case for Bug #26835. The bug could be reproduced as follows:
Define a table so that the first column of the clustered index is
a VARCHAR or a UTF-8 CHAR in a collation where sequences of bytes
of differing length are considered equivalent.

Insert and delete a record.  Before the delete-marked record is
purged, insert another record whose first column is of different
length but equivalent to the first record.  Under certain conditions,
the insertion can be incorrectly performed as update-in-place.

Likewise, an operation that could be done as update-in-place can
unnecessarily be performed as delete and insert, but that would not
cause corruption but merely degraded performance.

--BZR--
revision-id: svn-v4:cee13dc7-1704-0410-992b-c9b4543f1246:trunk:1340
testament3-sha1: 7e84689a80057f11ae8394b3516da1940e0058c3