Makefile.am     In the "test-bt" target, replace the call "$(MAKE) test-embedded" by the actions of that target.
--BZR--
revision-id: sp1r-joerg@trift2.-20071207105459-39181
property-file-info: ld7:file_id64:sp1f-makefile.am-19700101030959-jbbpiygwpgybyqknlavdxxupbrjonu7h7:message537:In the "test-bt" target, replace the call "$(MAKE) test-embedded" by the actions of that target.
property-file-info: 
property-file-info: Reason:
property-file-info: For release builds, we run the tests in a tree that contains just a binary installation.
property-file-info: So the Makefile is not in the test tree, and we explicitly address it in the build tree.
property-file-info: This explicit path is not kept by "make", so it is missing in the sub-make.
property-file-info: There seems to be no way to solve this with (portable!) "make" features -
property-file-info: we could add a new variable and use it in the sub-make, but it is easier to duplicate these few lines.
property-file-info: 4:path11:Makefile.amee
property-sp1-file-info: ld9:commit_id46:joerg@trift2.|Makefile.am|20071207105455|178717:file_id52:BK|Makefile.am|19700101030959|00030|ca05e0a7eadf0c15ee
testament3-sha1: 723ca7560dea53fc2efdbf5f43cf52b2782a9266