branches/5.1: Port of r2267
This is a combination of changes that forward port the scalability fix applied to 5.0
through r1001.
It reverts changes r149 and r122 (these were 5.1 specific changes made in lieu of
scalability fix of 5.0)
Then it applies r1001 to 5.0 which is the original scalability fix.
Finally it applies r2082 which fixes an issue with the original fix.

Reviewed by: Heikki

--BZR--
revision-id: svn-v4:cee13dc7-1704-0410-992b-c9b4543f1246:branches/5.1:2268
testament3-sha1: 162eaabaa5ac78b58908d4fe8a2ef3bd002fccce