branches/zip: fix Mantis issue #74 Memory leak on Windows
The memory leak was due to wrong parameters passed into VirtualFree()
call. So, the call fails with Windows error 87. MEM_DECOMMIT can NOT be
used along with MEM_RELEASE. And if the parameter is MEM_RELEASE, the
size parameter must be 0. The function frees the entire region that is
reserved in the initial allocation call to VirtualAlloc.

This issue was introduced by r984.

Approved by:	Heikki (on IM)
--BZR--
revision-id: svn-v4:16c675df-0fcb-4bc9-8058-dcc011a37293:branches/zip:2657
testament3-sha1: 1d103fe4ced2521bbd1b062a088fe4066744ed49