Bug#11122: Server won't always start when cold-booting after a crash
The grep expression that finds a running "mysqld" program fails if the
"mysqld_safe" is running with the same PID.

Now, excise "ps" output that has the word " grep" or "mysqld_safe" in 
it, to be a little more certain that the matched process is not a false 
positive hit.  This will fail when the path to mysqld contains either
of those two names, which should be acceptable.

Additionally, some text to search could be truncated if very long.  
Expand the number of lines "ps" emits.

--BZR--
revision-id: chad@mysql.com-20081002162552-cw77j2cpzw23qycy
property-branch-nick: 5.0-bugteam--bug11122
testament3-sha1: 8a6a7102a31952100d22f685e7e2ea071e694150