Bug#36768 (partition_info::check_partition_info() reports mal formed warnings)
Before this fix, several places in the code would raise a warning with an
error code 0, making it impossible for a stored procedure, a connector,
or a client application to trigger logic to handle the warning.
Also, the warning text was hard coded, and therefore not translated.

With this fix, new errors numbers have been created to represent these
warnings, and the warning text is coded in the errmsg.txt file.

--BZR--
revision-id: marc.alff@sun.com-20081006203615-srnz8p6nl2ljm2wl
property-branch-nick: mysql-5.1-36768
testament3-sha1: ab3542845908f76deacd56f690bfbfc8fe4d768d