Bug#39178: non-RSA keys in connection to a RSA-keyed yaSSL-using server \ 		using crashes server
When the server is configured to use a RSA key, and when the client sends
a cipher-suite list that contains a non-RSA key as acceptable, the server 
would try to process that key even though it was impossible.

Now, yaSSL sets its own acceptable-cipher list according to what kind of
key the server is started with, and will never explore and try to pair 
impossible combinations.

This involves a partial import of the current YaSSL tree, not the whole
thing, so as to try to avoid introducing new bugs.

(Updated to avoid many whitespace changes and make diff smaller.)

--BZR--
revision-id: chad@mysql.com-20081118164544-wz4nphj48egf474p
property-branch-nick: mysql-5.1-bugteam--bug39178
testament3-sha1: 27b40823119ff167b0bc90c6a4a353f94ae140d1