Bug#41002 - symlink.test fails on symlinked datadir
symlink.test failed when run in an environment that has mysql-test/var
symlinked to elsewhere, e.g. a memory file system. This is the case
when running mysql-test-run --mem.

In this case the server does not detect that the directory specified
with a DATA/INDEX DIRECTORY clause is within its data home directory.
This problem was reported as Bug#39277 (Creation of table with data
and/or index files in data home directory succeeds). It was decided
that it will not be fixed in 5.1. Hence, the current behavior is
accepted for 5.1. It will be fixed in 6.0 though.

Fixed the test case so that it works in both environments. 1. When no
symbolic link is involved, the server notices that the data/index
directory is in its data hone directory and rejects the CREATE/ALTER
TABLE statement. 2. When the data home directory is symlinked, it
does not notice the problem and executes the statement sucessfully.

--BZR--
revision-id: ingo.struewing@sun.com-20081125140631-qqnfima1u8ie5778
property-branch-nick: mysql-5.1-bug41002-1
property-file-info: ld7:file_id67:sp1f-symlink.result-20010605003810-qspjjr4hwq7gpdhzyictx52ddrlmxaco7:message72:Bug#41002 - symlink.test fails on symlinked datadir
property-file-info: Updated test result.4:path27:mysql-test/r/symlink.resulted7:file_id65:sp1f-symlink.test-20010605003810-7slktyzeszzfljtnmdbque4ow3azjqty7:message160:Bug#41002 - symlink.test fails on symlinked datadir
property-file-info: Adjusted the test case to the accepted behavior.
property-file-info: It needs to accept success and failure of some statements.
property-file-info: 4:path25:mysql-test/t/symlink.testee
testament3-sha1: 6c392460faea87428d2ecc2f260fed7490addeb2