Bug #40129  no-threads fails on pushbuild in 6.0-rpl, ps_stm_threadpool
The test reacted on the way how mtr orders arguments for the server
that are gathered from different source. It appeared that the opt-file
options were parsed before those that supplied to mtr via its command
line. In effect, the opt-file preferences got overriden by the command
line and some tests, like no-threads, were caught by surprise: a test
expects an option value that had been "hardcoded" into its opt-file
but gets another one.
This server options ordering problem exists on in the new rpl trees
mtr.  In option of the author of this patch, the opt-file shall be
considered as having the highest preference weight. The opt-file is
merely a part of the header of a test, namely a part that can not be
technically deployed along the test file.
It's unnatural for the test writer to provide both the opt file value
and a guard that guarantees the value will be set on in the run time.
It's logical to provide either one: the option and its value or the
guard.

Fixed with relocating parse of the opt file to be the last among
sources of the sever's options.

A side effect: fixing a small problem of resetting the suite options
at time the opt file starts parsing.

A side effect: main.log_bin_trust_function_creators_func is disabled to
be re-enabled with the fixes for bug#41003 will be merged from the main trees.

--BZR--
revision-id: aelkin@mysql.com-20081212215144-r0wvw3pz9f2pstbe
property-branch-nick: 5.1-rpl-bug41003-40129-log_bin_trust_no-threads
property-file-info: ld7:file_id65:sp1f-mtr_cases.pl-20050203205008-rrteoawyobvgq6u7zeyce4tmuu334ayg7:message257:Relocating parse of the opt file to be the last. This ensure the opt file is the last
property-file-info: provider for the server options so that the opt-file options have the highest preference;
property-file-info: fixing a separate issue of incorrect resetting the suite options for the server;
property-file-info: 4:path27:mysql-test/lib/mtr_cases.pmed7:file_id65:sp1f-disabled.def-20050315184020-inpdp4hiogithilv62snllppjz2dcing7:message126:log_bin_trust_function_creators_func is disabled. Todo: to-reable when fixes for bug#41003
property-file-info: will be merged from the main trees.4:path25:mysql-test/t/disabled.defee
testament3-sha1: 22a2fb7b2ed6b84fddc15ea3ceb2d5a56a102048