Apply InnoDB snapshot innodb-5.1-ss2858, part 5.
A follow-up to the previous patch, fix for Bug #39438.

r2720 | vasil | 2008-10-03 19:52:39 +0300 (Fri, 03 Oct 2008) | 8 lines
branches/5.1:

Print a warning if an attempt is made to get the free space for a table
whose .ibd file is missing or the tablespace has been discarded. This is a
followup to r2719.

Suggested by:	Inaam

--BZR--
revision-id: timothy.smith@sun.com-20081214203448-lnupcywat6sl5wub
property-branch-nick: 5.1-bugteam
testament3-sha1: 443b0155afeae7258dcafb1da09b4763d8073a53