Apply InnoDB snapshot innodb-5.1-ss2858, part 6.
Send error message to the client when a rollback occurs when trying to get an
autoincrement value.  This is related to the fixes for Bug#36411 and
Bug#35498.

Detailed revision comments:

r2721 | sunny | 2008-10-04 02:08:23 +0300 (Sat, 04 Oct 2008) | 6 lines
branches/5.1: We need to send the messages to the client because
handler::get_auto_increment() doesn't allow a way to return the
specific error for why it failed.

rb://18

--BZR--
revision-id: timothy.smith@sun.com-20081214203814-on1dzn3nwua4ylp3
property-branch-nick: 5.1-bugteam
testament3-sha1: fa2f2b01b0ad04ad5d66108caae5782b016128c9