Bug#41889: Test main.innodb_bug38231 is failing w/embedded server (server crash)
The problem is that a mysql connection instance is not thread-safe
and reentrant, meaning that it can't be used concurrently and can't
be re-entered while it's already running. This applies for any form
of the server (embedded or not), but this rule can be violated in a
test case if the test sends a new command without waiting for the
result of previous command that was sent asynchronously and this can
lead to hangs when over a network or to crashes under embedded server
as the server query execution path will be re-entered concurrently
with the same connection structure.

The solution is to rework the test case so that the aforementioned
rule is obeyed.
--BZR--
revision-id: davi.arnaut@sun.com-20090108123134-x204djl2e83ra391
property-branch-nick: 41889-5.1
property-file-info: ld7:file_id54:innodb_bug38231.test-20081214200020-x6z2ierhzp8f40y4-27:message240:Remove con3 as it is not necessary to reproduce the test case
property-file-info: and might cause problems as there is no guarantee on which
property-file-info: LOCK TABLE request will succeed first. Also, wait for statement
property-file-info: result before sending a new one on the same connection.4:path33:mysql-test/t/innodb_bug38231.testee
testament3-sha1: 339ef9d65d72e583b45d5480894989b7bf78e63e